% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids}
\alias{centroids}
\title{Class centroid locations}
\format{
An object of class \code{data.frame} with 1942 rows and 6 columns.
}
\usage{
centroids
}
\description{
A dataset containing the location of the four class centroids across 1942
genes, along with the HUGO and Ensembl gene identifiers for each.
}
\keyword{datasets}
