% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyNMIBC.R
\name{classifyNMIBC}
\alias{classifyNMIBC}
\title{Transcriptomic UROMOL2021 classes of non-muscle-invasive bladder cancer
(NMIBC) samples}
\usage{
classifyNMIBC(
  x,
  min_cor = 0.2,
  gene_id = c("ensembl_gene_ID", "hgnc_symbol"),
  tidy = FALSE
)
}
\arguments{
\item{x}{\code{data.frame} with unique genes as rows and samples as columns.
RNA-seq data needs to be log-transformed and micro-array data should be
normalized.}

\item{min_cor}{Numeric specifying a minimal threshold for best Pearson's
correlation between a sample's gene expression profile and centroids
profiles. A sample showing no correlation above this threshold will remain
unclassifed and prediction results will be set to NA.}

\item{gene_id}{Character specifying the type of gene identifiers used for the
rownames or first column of \code{x}}

\item{tidy}{Logical. If TRUE, assumes the first column contains the gene
identifiers. Otherwise, assumes IDs are rownames}
}
\value{
A \code{tibble} containing:
\describe{
\item{estimate}{Pearson correlation of a given \code{sample} to the given \code{class} centroid}
\item{conf.low, conf.high}{low and high end of 95\% confidence interval}
\item{nearest}{centroid for which the sample has the highest correlation to}
\item{statistic}{t-statistic}
\item{parameter}{degrees of freedom}
\item{sep_lvl}{(Highest Correlation - 2nd Highest Correlation)/median(Distance to highest correlation)}
}
}
\description{
This package performs Pearson nearest-centroid classification according to
the transcriptomic classes of NMIBC based on gene expression profiles.
}
\note{
This classifier was built similarly to the published tool for the
consensus classes of MIBC: Kamoun, A et. al. A Consensus Molecular
Classification of Muscle-invasive Bladder Cancer. Eur Uro (2019), doi:
https://doi.org/10.1016/j.eururo.2019.09.006
}
\examples{
data(test_data)
classifyNMIBC(test_data)

}
