% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data}
\format{
An object of class \code{data.frame} with 38990 rows and 20 columns.
}
\usage{
test_data
}
\description{
Pre-normalized test data, where columns are samples and rows are genes. Rownames are Ensembl IDs.
}
\keyword{datasets}
