% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyNMIBC.R
\name{classifyNMIBC}
\alias{classifyNMIBC}
\title{Transcriptomic UROMOL2021 classes of non-muscle-invasive bladder cancer (NMIBC) samples}
\usage{
classifyNMIBC(
  x,
  minCor = 0.2,
  gene_id = c("ensembl_gene_ID", "hgnc_symbol")[1]
)
}
\arguments{
\item{x}{dataframe with unique genes in rows and samples to be classified in columns (or single named vector of gene expression values).
RNA-seq data needs to be log-transformed and micro-array data should be normalized. Gene names may be supplied as Ensembl gene IDs or HUGO gene symbols.}

\item{minCor}{a numeric value specifying a minimal threshold for best Pearson's correlation between a sample's gene expression profile and centroids profiles.
A sample showing no correlation above this threshold will remain unclassifed and prediction results will be set to NA. Default minCor value is 0.2.}

\item{gene_id}{a character value specifying the type of gene identifiers used for the names/rownames of 'x',
ensembl_gene_ID for Ensembl gene IDs or hgnc_symbol for HUGO gene symbols.}
}
\value{
A dataframe with classification results. The NMIBC_Class column returns the predicted class labels of the samples.
The cor_pval column returns the p-value associated with the Pearson's correlation between the sample and the nearest centroid.
The separationLevel gives a measure (ranging from 0 to 1) of how a sample is representative of its consensus class,
with 0 meaning the sample is too close to the other classes to be confidently assigned to one class label, and 1 meaning the sample is very representative of its class.
This separationLevel is measured as follows: (correlation to nearest centroid - correlation to second nearest centroid) / median difference of sample-to-centroid correlation.
The remaing four columns return the Pearson's correlation values for each sample and each centroid.
NMIBC_Class predictions are set to NA if the minCor condition is not verified.
}
\description{
This package performs Pearson nearest-centroid classification according to the transcriptomic classes of NMIBC based on gene expression profiles.
}
\note{
This classifier was built similarly to the published tool for the consensus classes of MIBC:
Kamoun, A et. al. A Consensus Molecular Classification of Muscle-invasive Bladder Cancer. Eur Uro (2019), doi: https://doi.org/10.1016/j.eururo.2019.09.006

Alternatively, you can classifiy samples through our online web application: http://nmibc-class.dk
}
\examples{
data(test_data)
classifyNMIBC(test_data)

}
